refac: moved "is used" checks into it's own class to reduce class complexity
